@@ -5,87 +5,110 @@
   "main": "index.js",
   "scripts": {
     "start": "cd build && node main",
-    "start:server": "node main | pino -o 'short'",
+    "start:server": "node main | pino-pretty",
     "compile:local": "node build-scripts/build.js && backpack build",
     "compile:dev": "node build-scripts/build.dev.js && backpack build",
     "compile:test": "node build-scripts/build.test.js && backpack build",
     "compile:qa": "node build-scripts/build.qa.js && backpack build",
     "compile:prod": "node build-scripts/build.prod.js && backpack build",
-    "serve": "node build-scripts/build.js && backpack dev | pino -o 'short'",
-    "serve:dev": "node build-scripts/build.dev.js && backpack dev | pino -o 'short'",
-    "serve:test": "node build-scripts/build.test.js && backpack dev | pino -o 'short'",
-    "serve:qa": "node build-scripts/build.qa.js && backpack dev | pino -o 'short'",
-    "serve:prod": "node build-scripts/build.prod.js && backpack dev | pino -o 'short'",
+    "serve": "node build-scripts/build.js && backpack dev | pino-pretty",
+    "serve:debug": "node build-scripts/build.js && backpack dev --inspect | pino-pretty",
+    "serve:dev": "node build-scripts/build.dev.js && backpack dev | pino-pretty",
+    "serve:test": "node build-scripts/build.test.js && backpack dev | pino-pretty",
+    "serve:qa": "node build-scripts/build.qa.js && backpack dev | pino-pretty",
+    "serve:prod": "node build-scripts/build.prod.js && backpack dev | pino-pretty",
     "test": "node build-scripts/build.js && mocha test/**/*.ts --exit",
     "test:watch": "node build-scripts/build.js && mocha -r ts-node/register mocha test/**/*.ts  --watch --watch-extensions ts",
     "test:report": "node build-scripts/build.js && mocha test/**/*.ts --reporter mochawesome --exit",
     "test:coverage": "nyc mocha test/**/*.ts --exit --require ts-node/register --require source-map-support/register --recursive"
   },
   "dependencies": {
-    "body-parser": "^1.18.2",
-    "class-validator": "^0.9.1",
-    "cookie-parser": "^1.4.3",
+    "@typegoose/typegoose": "^6.4.0",
+    "body-parser": "^1.19.0",
+    "class-validator": "^0.11.0",
+    "cookie-parser": "^1.4.4",
     "cookie-session": "^2.0.0-beta.3",
-    "dotenv": "^4.0.0",
-    "express": "^4.16.2",
-    "express-validator": "^5.0.3",
-    "httpntlm": "^1.7.5",
-    "install": "^0.11.0",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "express-validator": "^6.4.0",
+    "inversify": "^5.0.1",
+    "inversify-express-utils": "^6.3.2",
     "istanbul": "^0.4.5",
-    "js-yaml": "^3.12.0",
-    "json-refs": "^2.1.6",
-    "json2typescript": "^1.0.5",
-    "jsonwebtoken": "^8.2.1",
-    "jwt-simple": "^0.5.1",
-    "lodash": "^4.17.10",
-    "mongoose": "^5.0.11",
-    "nock": "^9.2.5",
-    "openid-client": "^2.0.1",
-    "passport": "^0.4.0",
-    "pino": "^4.7.3",
-    "portscanner": "^2.1.1",
-    "reflect-metadata": "^0.1.12",
-    "request": "^2.85.0",
-    "request-promise": "^4.2.2",
-    "source-map-support": "^0.4.15",
-    "supertest-session": "^3.2.0",
-    "swagger-express-middleware": "^1.0.0-alpha.12",
-    "typegoose": "^5.0.0",
-    "types-installer": "^1.4.0"
+    "js-yaml": "^3.13.1",
+    "json-refs": "^3.0.15",
+    "json2typescript": "^1.2.5",
+    "mongoose": "^5.9.4",
+    "nock": "^12.0.2",
+    "passport": "^0.4.1",
+    "pino": "^5.17.0",
+    "pino-pretty": "^3.6.1",
+    "portscanner": "^2.2.0",
+    "prettyjson": "^1.2.1",
+    "reflect-metadata": "^0.1.13",
+    "request": "^2.88.2",
+    "request-promise": "^4.2.5",
+    "source-map-support": "^0.5.16",
+    "swagger-express-middleware": "^2.0.5"
   },
   "devDependencies": {
-    "@types/body-parser": "^1.16.8",
-    "@types/chai": "^4.1.2",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/cookie-session": "^2.0.34",
-    "@types/dotenv": "^4.0.2",
-    "@types/express": "^4.11.1",
-    "@types/jsonwebtoken": "^7.2.7",
+    "@types/body-parser": "^1.19.0",
+    "@types/chai": "^4.2.10",
+    "@types/cookie-parser": "^1.4.2",
+    "@types/cookie-session": "^2.0.38",
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.3",
+    "@types/jsonwebtoken": "^8.3.8",
     "@types/jwt-simple": "^0.5.33",
-    "@types/lodash": "^4.14.108",
-    "@types/mocha": "^5.0.0",
-    "@types/mongoose": "^5.0.7",
-    "@types/passport": "^0.4.5",
-    "@types/pino": "^4.7.1",
+    "@types/lodash": "^4.14.149",
+    "@types/mocha": "^7.0.2",
+    "@types/mongoose": "^5.7.5",
+    "@types/passport": "^1.0.3",
+    "@types/pino": "^5.17.0",
     "@types/portscanner": "^2.1.0",
-    "@types/request-promise": "^4.1.41",
-    "@types/shelljs": "^0.7.8",
-    "@types/source-map-support": "^0.4.0",
-    "@types/supertest": "^2.0.4",
-    "@types/swagger-express-middleware": "^1.0.8",
-    "awesome-typescript-loader": "^3.4.1",
-    "backpack-core": "^0.4.3",
-    "chai": "^4.1.2",
-    "mocha": "^4.0.1",
-    "mochawesome": "^3.0.2",
-    "nyc": "^11.6.0",
-    "shelljs": "^0.7.8",
-    "supertest": "^3.0.0",
-    "ts-node": "^5.0.1",
-    "tsconfig-paths": "^3.5.0",
+    "@types/request-promise": "^4.1.45",
+    "@types/shelljs": "^0.8.6",
+    "@types/source-map-support": "^0.5.1",
+    "@types/supertest": "^2.0.8",
+    "@types/swagger-express-middleware": "^1.0.10",
+    "awesome-typescript-loader": "^5.2.1",
+    "babel-polyfill": "^6.26.0",
+    "backpack-core": "^0.8.4",
+    "base64-js": "^1.3.1",
+    "chai": "^4.2.0",
+    "commander": "^4.1.1",
+    "deep-extend": "^0.6.0",
+    "ieee754": "^1.1.13",
+    "immutable": "^4.0.0-rc.12",
+    "isarray": "^2.0.5",
+    "memoizee": "^0.4.14",
+    "mocha": "^7.1.0",
+    "mochawesome": "^5.0.0",
+    "nyc": "^15.0.0",
+    "react": "^16.13.0",
+    "react-addons-shallow-compare": "^15.6.2",
+    "react-collapse": "^5.0.1",
+    "react-dom": "^16.13.0",
+    "react-immutable-proptypes": "^2.1.0",
+    "react-redux": "^7.2.0",
+    "react-remarkable": "^1.1.3",
+    "react-split-pane": "^0.1.89",
+    "redux": "^4.0.5",
+    "redux-immutable": "^4.0.0",
+    "reselect": "^4.0.0",
+    "sanitize-html": "^1.22.0",
+    "serialize-error": "^5.0.0",
+    "shallowequal": "^1.1.0",
+    "shelljs": "^0.8.3",
+    "supertest": "^4.0.2",
+    "swagger-client": "^3.10.0",
+    "ts-node": "^8.6.2",
+    "tsconfig-paths": "^3.9.0",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "tslint": "^5.3.2",
-    "typescript": "^3.0.1"
+    "tslint": "^6.0.0",
+    "typescript": "^3.8.3",
+    "url-parse": "^1.4.7",
+    "xml": "^1.0.1",
+    "yaml-js": "^0.2.3"
   },
   "nyc": {
     "extension": [
