@@ -1,14 +1,19 @@
+/* -------------------------------------------------------------------------- */
+/*                                   IMPORTS                                  */
+/* -------------------------------------------------------------------------- */
+/* ------------------------------- THIRD PARTY ------------------------------ */
 import * as middleware from "swagger-express-middleware";
 import { Application } from "express";
 import * as path from "path";
-import { logger } from "../logger";
-const resolve = require("json-refs").resolveRefs;
-const YAML = require("js-yaml");
-const fs = require("fs");
+/* --------------------------------- CUSTOM --------------------------------- */
+import { routes } from "../../api/routes";
 
 
-export const swaggerify = function (app: Application, routes: (app: Application) => void) {
-  middleware(path.join(__dirname, "Api.yaml"), app, function (err, middleware) {
+/* -------------------------------------------------------------------------- */
+/*                                   EXPORTS                                  */
+/* -------------------------------------------------------------------------- */
+export const swaggerify = function (app: Application) {
+  middleware(path.join(__dirname, "Api.yaml"), app, (err, middleware) => {
     app.use(middleware.files(app, {
       apiPath: process.env.SWAGGER_API_SPEC,
     }));
