@@ -4,6 +4,7 @@ import { wordArt } from "./WordArt";
 export interface IOutputDirectories {
     app: {
         base: string;
+        injectables: string
     }
     models: {
         base: string;
@@ -34,7 +35,7 @@ export interface IOutputDirectories {
 class Constants {
     GENERATOR_NAME: string = 'wilbur';
 
-    // Filepaths
+    // File paths
     filePaths = {
         templates: {
             base: 'templates',
@@ -53,12 +54,14 @@ class Constants {
             },
             controller: {
                 base: '/controller',
+                inversify : '/controller/inversify/class.controller.ejs',
                 controller: '/controller/class.controller.ejs',
                 router: '/controller/class.router.ejs',
                 validator : '/controller/class.validator.ejs'
             },
             service : {
                 base : '/service',
+                inversify : '/service/inversify/class.service.ejs',
                 service : '/service/class.service.ejs'
             },
             other : {
@@ -72,6 +75,7 @@ class Constants {
         return {
             app: {
                 base: baseOutputDirectory,
+                injectables : baseOutputDirectory + '/src/common/inversify/injectables.ts'
             },
             models: {
                 base: baseOutputDirectory + '/src/models',
@@ -91,7 +95,7 @@ class Constants {
             controller: {
                 base: baseOutputDirectory + '/src/api/controllers',
                 index: baseOutputDirectory + '/src/api/controllers/index.ts',
-                routes : baseOutputDirectory + '/src/routes.ts'
+                routes : baseOutputDirectory + '/src/api/routes.ts'
             },
             service : {
                 base : baseOutputDirectory + '/src/api/services',
