@@ -12,12 +12,14 @@ export class ServiceService extends Generator {
     tsClass: Class;                                 // Class to be generated/removed
     fileUtil = new FileUtils();                     // Provides common filesystem functionalities
     private classNameCamelCase: string;             // Camel case name of class to be generated/removed
+    useDI = true;                                   // Use dependency injection
 
 
     constructor(args: any, options: any, private appName: string) {
         super(args, options);
         this.generalUtils = new GeneralUtils(args, options);
         this.outputDirectories = constants.generateOutputDirectories(this.destinationPath(this.appName));
+        this.useDI = (typeof options.useDI === 'boolean') ? options.useDI : true;
     }
 
     /* Given a class, generate all necessary service files to existing application */
@@ -25,7 +27,8 @@ export class ServiceService extends Generator {
         this.tsClass = tsClass;                                                                                 // Read existing file contents
         this.classNameCamelCase = this.tsClass.name.charAt(0).toLowerCase() + this.tsClass.name.substring(1);   // Name of class to add/remove
         this.generateServiceFile();                                                                             // Generate service file
-        this.updateServiceIndexFile(ClassActionType.ADD_EDIT);                                                  // Modify services index file   
+        this.updateServiceIndexFile(ClassActionType.ADD_EDIT);                                                  // Modify services index file
+        this.updateDIContainer(ClassActionType.ADD_EDIT);                                                       // Modify DI container
     }
 
     /* Given a class, remove service file and its references */
@@ -34,6 +37,7 @@ export class ServiceService extends Generator {
         this.classNameCamelCase = this.tsClass.name.charAt(0).toLowerCase() + this.tsClass.name.substring(1);   // Name of class to add/remove
         await this.removeServiceFiles();                                                                        // Delete service file from filesystem
         this.updateServiceIndexFile(ClassActionType.REMOVE);                                                    // Remove reference to service files
+        this.updateDIContainer(ClassActionType.REMOVE);                                                         // Modify DI container
     }
 
     /* Delete service file from filesystem */
@@ -44,8 +48,10 @@ export class ServiceService extends Generator {
 
     /* Generate a service file from a template */
     private generateServiceFile() {
+        const templatePath = this.generalUtils.directories.templates.service;
+
         this.fs.copyTpl(
-            this.generalUtils.directories.templates.service.service,
+            (this.useDI) ? templatePath.inversify : templatePath.service,
             this.outputDirectories.service.base + "/" + this.classNameCamelCase + '/' + this.classNameCamelCase + ".service.ts",
             { tsClass: this.tsClass }
         );
@@ -60,10 +66,40 @@ export class ServiceService extends Generator {
             if (fileContents.indexOf(importContent) === -1)                                                     // Search if content already exists in file
                 fileContents += `\n${importContent}`;                                                           // Add content to file
         } else {
-            fileContents = fileContents.replace(importContent, "");                                             // Add content to file
+            fileContents = fileContents.replace(importContent, "");                                             // Remove import content from file
         }
 
         this.fs.write(this.outputDirectories.service.index, fileContents);                                      // Add content to file
+    }
+
+    private updateDIContainer(actionType: ClassActionType) {
+        console.log('sup')
+        if (!this.useDI) return
+        console.log('nmu')
+        let fileContents = this.fs.read(this.outputDirectories.app.injectables);                                // Read existing file contents               
+        const serviceClassName = `${this.tsClass.name}Service`,
+            importContent = `import { ${serviceClassName} } from "@services";`,
+            containerDeclaration = `container.bind<${serviceClassName}>("${serviceClassName}").to(${serviceClassName});`,
+            importComment = "// SERVICE IMPORTS - **DO NOT REMOVE THIS COMMENT**",
+            registerFunction = "export const registerInjectables = (container: Container) => {";
+
+
+        if (actionType === ClassActionType.ADD_EDIT) {
+            if (fileContents.indexOf(importContent) === -1) {                                                   // Add service import
+                const insertionIndex = fileContents.indexOf(importComment) + importComment.length;
+                fileContents = fileContents.slice(0, insertionIndex) + `\n${importContent}` + fileContents.slice(insertionIndex)
+            }
+
+            if (!fileContents.includes(containerDeclaration)) {
+                const insertionIndex = fileContents.indexOf(registerFunction) + registerFunction.length;
+                fileContents = fileContents.slice(0, insertionIndex) + `\n\t${containerDeclaration}` + fileContents.slice(insertionIndex)
+            }
+        } else {
+            fileContents = fileContents.replace(importContent, "");                                             // Remove import content from file
+            fileContents = fileContents.replace(containerDeclaration, "");                                             // Remove container declaration from file
+        }
+
+        this.fs.write(this.outputDirectories.app.injectables, fileContents);                                      // Add content to file
     }
 
 }
