@@ -1,9 +1,5 @@
-// GENERAL
+/* --------------------------------- GENERAL -------------------------------- */
 export * from "./general/general.controller";
-export * from "./general/general.router";
 
-// API-ROOT
+/* -------------------------------- API ROOT -------------------------------- */
 export * from "./api-root/api-root.controller";
-export * from "./api-root/api-root.router";
-
-
