@@ -1,16 +1,19 @@
-/*--------------------THIRD PARTY-------------------*/
-// import { ObjectId } from "bson";
-import * as request from "request-promise";
+/* -------------------------------------------------------------------------- */
+/*                                   IMPORTS                                  */
+/* -------------------------------------------------------------------------- */
+/* ------------------------------- THIRD PARTY ------------------------------ */
 import { JsonConvert, ValueCheckingMode } from "json2typescript";
-import { logger } from "@common";
-/*--------------------CUSTOM-------------------*/
-/*--------------------FUNCTIONS------------------------------------*/
-// create the options for a request
+/* --------------------------------- CUSTOM --------------------------------- */
+import { logAndThrowError } from "@common";
 
 
-/*--------------------CLASS-------------------*/
+/* -------------------------------------------------------------------------- */
+/*                              CLASS DEFINITION                              */
+/* -------------------------------------------------------------------------- */
 export class CommonService {
 
+
+    /* ---------------------------- MEMBER FUNCTIONS ---------------------------- */
     createHttpOptions(requestType: string, path: string, auth: string): any {
         return {
             method: requestType,
@@ -20,32 +23,34 @@ export class CommonService {
             }
         };
     }
-    // deserialize a json object into an object of type T
-    deserialize<T>(jsonObj: Object, type: any) {
+
+    withError = (promise: Promise<any>) => promise.then(data => [null, data]).catch(err => [err]);
+
+
+    /* ----------- Deserialize a json object into an object of type T ----------- */
+    deserialize<T>(jsonObj: Object, type: any): T {
         const jsonConvert: JsonConvert = new JsonConvert();
         jsonConvert.ignorePrimitiveChecks = false; // don"t allow assigning number to string etc.
         jsonConvert.valueCheckingMode = ValueCheckingMode.ALLOW_NULL; // allow null
-        let deserializedObj: T;
         try {
-            deserializedObj = jsonConvert.deserialize(jsonObj, type);
-            return deserializedObj;
+            return jsonConvert.deserializeObject(jsonObj, type);
         } catch (e) {
-            logger.error(<Error>e);
-            throw (e);
+            logAndThrowError(e);
         }
     }
-    // deserialize a json array object into an array of type T
+
+
+    /* --------- Deserialize a json array object into an array of type T -------- */
     deserializeArray<T>(jsonArray: Object[], type: any) {
         const jsonConvert: JsonConvert = new JsonConvert();
         jsonConvert.ignorePrimitiveChecks = false; // don"t allow assigning number to string etc.
         jsonConvert.valueCheckingMode = ValueCheckingMode.ALLOW_NULL; // allow null
         let deserializedArray: Array<T>;
         try {
-            deserializedArray = jsonConvert.deserializeArray(jsonArray, type);
+            deserializedArray = jsonConvert.deserializeArray<T>(jsonArray, type);
             return deserializedArray;
         } catch (e) {
-            logger.error(<Error>e);
-            throw (e);
+            logAndThrowError(e);
         }
     }
 }
