@@ -1,17 +1,29 @@
-/*--------------------THIRD PARTY-------------------*/
+/* -------------------------------------------------------------------------- */
+/*                                   IMPORTS                                  */
+/* -------------------------------------------------------------------------- */
+/* ------------------------------- THIRD PARTY ------------------------------ */
 import { Request, Response, NextFunction } from "express";
+import { Result, ValidationError, validationResult, check } from "express-validator";
+import { ServerResponse } from "http";
+import * as prettyjson from "prettyjson";
+/* --------------------------------- CUSTOM --------------------------------- */
 import { ABError, errorHandler } from "@models";
-/*--------------------CUSTOM-------------------*/
+import { logger } from "@common";
 
+/* -------------------------------------------------------------------------- */
+/*                             SERVICE DEFINITION                             */
+/* -------------------------------------------------------------------------- */
 class ControllerService {
-    /*--------------------CONSTRUCTOR----------------------------------*/
-    constructor() { }
-    /*--------------------FUNCTIONS------------------------------------*/
-    // parse errors and send them as a response
-    handleError(err: any, res: Response) {
+    /* ---------------------------- MEMBER VARIABLES ---------------------------- */
 
+    validateMongoID = [check("id", "Must include a valid MongoDB objectID in the request url.").isMongoId()];
+
+
+    /* --------------------------------- METHODS -------------------------------- */
+    // Parse errors and send them as a response
+    handleError(err: any, res: Response) {
         try {
-            // if error being handled has options or response, strip them out in order to hide api request credentials in error response body
+            // If error being handled has options or response, strip them out in order to hide api request credentials in error response body
             delete err["options"];
             delete err["headers"];
             delete err["response"];
@@ -23,9 +35,9 @@ class ControllerService {
                 err["message"] = errMsg;
             }
         }
-        catch (e) { /*couldnt parse error message, no big deal just continue*/ }
+        catch (e) { /*couldnt parse error message, just continue */ }
 
-        // de-serialize error and send it in a response body
+        // De-serialize error and send it in a response body
         if (!(err instanceof ABError))
             err = errorHandler.handleError(err);
         res.status(err.status).json(err);
@@ -39,15 +51,82 @@ class ControllerService {
             : false;
     }
 
-    assessErrors(req: Request, res: Response, next: NextFunction, errors: any, message?: string) {
+    assessErrors(req: Request, res: Response, next: NextFunction, errors: Result<ValidationError>, message?: string) {
         if (!message) message = "please make a valid request";
-        if (errors) {
+        if (!errors.isEmpty()) {
             const err = errorHandler.handleError(errors, `Bad request: ${message}`, 400);
             res.status(err.status).send(err);
         }
         else
             next();
     }
+
+    validateRequest(validations: any) {
+        return async (req: Request, res: Response, next: NextFunction) => {
+            await Promise.all(validations.map((validation: any) => validation.run(req)));
+
+            try {
+                validationResult(req).throw();
+                next();
+            } catch (errors) {
+                const err = errorHandler.handleError(errors.mapped(), "Bad request", 400);
+                res.status(err.status).send(err);
+            }
+        };
+    }
+
+    performRequest(fn: any) {
+        return async (req: Request, res: Response, next: NextFunction) => {
+            if (res.headersSent) return;
+
+            try {
+                const response = await fn(req, res, next);
+                // response has already been sent, return
+                if (response instanceof ServerResponse && response.headersSent) return;
+                return res.send(response);
+            } catch (err) {
+                logger.error(err);
+                if (err instanceof ServerResponse && err.headersSent) return;
+
+                try {
+                    // if the error message is in json, parse it instead of returning it as a string
+                    let errMsg = JSON.parse(JSON.stringify(err)).error;
+                    if (this.isJSONParsable(errMsg)) {
+                        errMsg = JSON.parse(errMsg);
+                        err.message = errMsg;
+                    }
+                }
+                catch (e) { /* Couldn't parse error message, just continue */ }
+
+                // de-serialize error and send it in a response body
+                if (!(err instanceof ABError))
+                    err = errorHandler.handleError(err);
+                return res.status(err.status).json(err);
+
+            }
+        };
+    }
+
+    requestErrorHandler = (err: any, req: Request, res: Response, next: NextFunction) => {
+        logger.error("\n" + prettyjson.render(err));
+        if (err instanceof ServerResponse && err.headersSent) return;
+
+        try {
+            // if the error message is in json, parse it instead of returning it as a string
+            let errMsg = JSON.parse(JSON.stringify(err)).error;
+            if (this.isJSONParsable(errMsg)) {
+                errMsg = JSON.parse(errMsg);
+                err.message = errMsg;
+            }
+        }
+        catch (e) { /* Couldn't parse error message, just continue */ }
+
+        // de-serialize error and send it in a response body
+        if (!(err instanceof ABError))
+            err = errorHandler.handleError(err);
+        return res.status(err.status).json(err);
+    };
+
 }
 
 // Exported Instance
