@@ -1,56 +1,80 @@
-/*--------------------THIRD PARTY-------------------*/
+/* -------------------------------------------------------------------------- */
+/*                                   IMPORTS                                  */
+/* -------------------------------------------------------------------------- */
+/* ------------------------------- THIRD PARTY ------------------------------ */
 import * as express from "express";
-import { Application } from "express";
+import { InversifyExpressServer, getRouteInfo } from "inversify-express-utils";
+import { Container } from "inversify";
 import * as path from "path";
-import * as expressValidator from "express-validator";
 import * as bodyParser from "body-parser";
 import * as http from "http";
 import * as os from "os";
 import * as cookieParser from "cookie-parser";
 import * as net from "net";
-import * as passport from "passport";
 import cookieSession = require("cookie-session");
-/*--------------------CUSTOM-------------------*/
-import { logger, swaggerify, mongoDatabase } from "@common";
-// import AuthService from "../api/services/auth/auth.service";
-/*------------------------------MODULE CONSTANTS---------------------*/
+// import * as passport from "passport";
+/* --------------------------------- CUSTOM --------------------------------- */
+import "@controllers";
+import { logger, swaggerify, mongoDatabase, createContainer } from "@common";
+import { controllerService } from "@services";
 
 
-/*------------------------------MODULE DEFINITION--------------------*/
-class ExpressServer {
-  app = express();
-  /*--------------------CONSTRUCTOR----------------------------------*/
+/* -------------------------------------------------------------------------- */
+/*                              CLASS DEFINITION                              */
+/* -------------------------------------------------------------------------- */
+export class ExpressServerNew {
+  /* ---------------------------- MEMBER VARIABLES ---------------------------- */
+  app: express.Application;
+  container: Container;
+
+  /* ------------------------------- CONSTRUCTOR ------------------------------ */
   constructor() {
-    const root = path.normalize(__dirname + "/../..");
-    this.app.set("this.appPath", root + "client");
-    this.app.use(expressValidator());
-    this.app.use(bodyParser.json());
-    this.app.use(bodyParser.urlencoded({ extended: true }));
-    this.app.use(cookieParser(process.env.SESSION_SECRET));
-    // this.app.use(express.static(`${root}/public`));
-    this.app.use("/docs", express.static(`${root}/public/api-explorer`));
-    process.env.NODE_TLS_REJECT_UNAUTHORIZED = "0";
-
-    this.app.use(cookieSession({
-      name: "asdf",
-      secret: "Asdf",
-      maxAge: 24 * 60 * 60 * 1000 // 24 hours
-    }));
-
-    // AuthService.setupPassport();
-    this.app.use(passport.initialize());
-    this.app.use(passport.initialize());
-    this.app.use(passport.session());
-
-    // this.app.use((req, res, next) => {
-    //   res.locals.user = req.user;
-    //   next();
-    // });
-
     logger.info(`Starting Server in ${process.env.ENV_NAME} environment`);
+
+    this.container = createContainer();               // Create app container
+    const server = this.createServer(this.container); // Create server
+    this.app = server.build();                        // Create application
+    swaggerify(this.app);                             // Generate Swagger docs
   }
 
-  /*--------------------FUNCTIONS------------------------------------*/
+  /* --------------------------------- METHODS -------------------------------- */
+  private createServer(container: Container) {
+    // Create server
+    const server = new InversifyExpressServer(container);
+
+    // Configure Server
+    server.setConfig((app) => {
+      // File path configurations
+      const root = path.normalize(__dirname + "/../..");
+      app.set("appPath", root + "client");
+      app.use("/docs", express.static(`${root}/public/api-explorer`));
+
+      // Body & cookie parsers
+      app.use(bodyParser.json());
+      app.use(bodyParser.urlencoded({ extended: true }));
+      app.use(cookieParser(process.env.SESSION_SECRET));
+      app.use(cookieSession({
+        name: "asdf",
+        secret: "Asdf",
+        maxAge: 24 * 60 * 60 * 1000 // 24 hours
+      }));
+
+      // Disable reject unauthorized
+      process.env.NODE_TLS_REJECT_UNAUTHORIZED = "0";
+
+      // Passport
+      // AuthService.setupPassport();
+      // app.use(passport.initialize());
+      // app.use(passport.initialize());
+      // app.use(passport.session());
+    });
+
+    server.setErrorConfig((app) => app.use(controllerService.requestErrorHandler));
+
+    return server;
+  }
+
+  // Scan port number to tell if port is available
   isPortTaken(port: number): Promise<boolean> {
     return new Promise<boolean>((resolve, reject) => {
       const tester: any = net.createServer()
@@ -60,16 +84,11 @@ class ExpressServer {
     });
   }
 
-  // Registers this.app with router & swagger docs
-  router(routes: (app: Application) => void): ExpressServer {
-    swaggerify(this.app, routes);
-    return this;
-  }
 
-  // makes connection to db
+  // Make connection to db
   async connectToDB(): Promise<boolean> {
     try {
-      const connected: boolean = await mongoDatabase.connect();
+      await mongoDatabase.connect();
       return true;
     }
     catch (err) {
@@ -78,11 +97,10 @@ class ExpressServer {
     }
   }
 
-  listen(port = parseInt(process.env.PORT)): Application {
-    const welcome = (port: number) => () => logger.info(`Server up and running in ${process.env.ENV_NAME || "development"} environment @: ${os.hostname()} on port: ${port}}`);
+  listen(port = parseInt(process.env.PORT)): express.Application {
+    const welcome = (port: number) => () => logger.info(`Server up and running in ${process.env.ENV_NAME || "development"} environment @: ${os.hostname()} on port ${port}`);
     http.createServer(this.app).listen(port, welcome(port));
     return this.app;
   }
-}
 
-export const app = new ExpressServer();
+}
